Aplicar a validação de caracteres suportados antes de quantidade de caracteres permitidos - migração dos mocks para classes
