+++ SME.CDEP.TesteIntegracao/Acervo/Ao_fazer_manutencao_acervo.cs
@@ -282,7 +282,7 @@ namespace SME.CDEP.TesteIntegracao
             var servicoAcervo = GetServicoAcervo();
             var random = new Random();
             
-            var acervoInserir = GerarAcervo(TipoAcervo.Bibliografico).Generate();
+            var acervoInserir = AcervoMock.Instance.GerarAcervo(TipoAcervo.Bibliografico).Generate();
             acervoInserir.CreditosAutoresIds = new long []{1,2,3,4,5};
             acervoInserir.CoAutores = new List<CoAutor>() { new () { CreditoAutorId = random.Next(1,5), TipoAutoria = faker.Lorem.Word().Limite(15)}}.ToList();
             
@@ -309,7 +309,7 @@ namespace SME.CDEP.TesteIntegracao
             await InserirDadosBasicos();
             var servicoAcervo = GetServicoAcervo();
             
-            var acervoInserir = GerarAcervo(TipoAcervo.Bibliografico).Generate();
+            var acervoInserir = AcervoMock.Instance.GerarAcervo(TipoAcervo.Bibliografico).Generate();
             acervoInserir.CreditosAutoresIds = new long []{1,2,3,4,5};
             acervoInserir.CoAutores = Enumerable.Empty<CoAutor>();
             
@@ -334,7 +334,7 @@ namespace SME.CDEP.TesteIntegracao
             await InserirAcervo();
             var servicoAcervo = GetServicoAcervo();
             
-            var acervoInserir = GerarAcervo(TipoAcervo.Bibliografico).Generate();
+            var acervoInserir = AcervoMock.Instance.GerarAcervo(TipoAcervo.Bibliografico).Generate();
             acervoInserir.CreditosAutoresIds = new long []{1,2,3,4,5};
             acervoInserir.CoAutores = new List<CoAutor>() { new () { CreditoAutorId = new Random().Next(1,5), TipoAutoria = faker.Lorem.Word().Limite(15)}}.ToList();
 
@@ -353,7 +353,7 @@ namespace SME.CDEP.TesteIntegracao
             await InserirAcervo(TipoAcervo.DocumentacaoHistorica);
             var servicoAcervo = GetServicoAcervo();
             
-            var acervoInserir = GerarAcervo(TipoAcervo.Bibliografico).Generate();
+            var acervoInserir = AcervoMock.Instance.GerarAcervo(TipoAcervo.Bibliografico).Generate();
             acervoInserir.CreditosAutoresIds = new long []{1,2,3,4,5};
             acervoInserir.CoAutores = new List<CoAutor>() { new () { CreditoAutorId = new Random().Next(1,5), TipoAutoria = faker.Lorem.Word().Limite(15)}}.ToList();
