+++ SME.CDEP.TesteIntegracao/AcervoBibliografico/Ao_fazer_manutencao_acervo_bibliografico.cs
@@ -65,7 +65,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoAtual = ObterTodos<Acervo>().FirstOrDefault(w=> w.Id == 3);
             
@@ -142,7 +142,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoAtual = ObterTodos<Acervo>().FirstOrDefault(w=> w.Id == 3);
             
@@ -219,7 +219,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoAtual = ObterTodos<Acervo>().FirstOrDefault(w=> w.Id == 3);
             
@@ -246,7 +246,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoAtual = ObterTodos<Acervo>().FirstOrDefault(w=> w.Id == 3);
             
@@ -274,7 +274,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoAtual = ObterTodos<Acervo>().FirstOrDefault(w=> w.Id == 3);
             
@@ -346,7 +346,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoBibliograficoCadastroDto = mapper.Map<AcervoBibliograficoCadastroDTO>(acervoBibliograficoCompleto);
             acervoBibliograficoCadastroDto.CreditosAutoresIds = new long[] { 1, 2, 3 };
@@ -414,7 +414,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoBibliograficoCadastroDto = mapper.Map<AcervoBibliograficoCadastroDTO>(acervoBibliograficoCompleto);
             acervoBibliograficoCadastroDto.CreditosAutoresIds = new long[] { 4,5 };
@@ -482,7 +482,7 @@ namespace SME.CDEP.TesteIntegracao
             
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
             
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
             
             var acervoBibliograficoCadastroDto = mapper.Map<AcervoBibliograficoCadastroDTO>(acervoBibliograficoCompleto);
             acervoBibliograficoCadastroDto.CreditosAutoresIds = new long[] { 4,5 };
@@ -506,7 +506,7 @@ namespace SME.CDEP.TesteIntegracao
         
             var servicoAcervobibliografico = GetServicoAcervoBibliografico();
         
-            var acervoBibliograficoCompleto = GerarAcervoBibliografico().Generate();
+            var acervoBibliograficoCompleto = AcervoBibliograficoMock.Instance.GerarAcervoBibliografico().Generate();
         
             var acervoBibliograficoCadastroDto = mapper.Map<AcervoBibliograficoCadastroDTO>(acervoBibliograficoCompleto);
             acervoBibliograficoCadastroDto.CreditosAutoresIds = new long[] { 4,5 };
